Remove indexNumber from Worker. Add generators for workers and student. Move implementation of Commands to cpp file
