+++ Commands/Command.hpp
@@ -3,7 +3,6 @@
 #include <string>
 class Command {
 public:
-    Command() = default;
     virtual ~Command() = default;
 
     virtual void run() = 0;
