+++ Worker.hpp
@@ -5,7 +5,7 @@
 
 class Worker : public Person {
 public:
-    Worker(std::string firstName, std::string surName, std::string city, std::string street, std::string numberOfStreet, size_t indexNumber, std::string pesel, Gender gender, size_t income);
+    Worker(std::string firstName, std::string surName, std::string city, std::string street, std::string numberOfStreet, std::string pesel, Gender gender, size_t income);
     ~Worker() override = default;
 
     void setFirstName(const std::string& firstName) override;
@@ -13,7 +13,6 @@ public:
     void setCity(const std::string& city) override;
     void setStreet(const std::string& street) override;
     void setNumberOfStreet(const std::string& numberOfStreet) override;
-    void setIndexNumber(const size_t& indexNumber) override;
     void setPesel(const std::string& pesel) override;
     void setGender(const Gender& gender) override;
     void setIncome(const size_t income) override { income_ = income; }
@@ -21,12 +20,11 @@ public:
     const std::string& getFirstName() const override;
     const std::string& getSurName() const override;
     const std::string getAddress() const override;
-    size_t getIndexNumber() const override;
     const std::string& getPesel() const override;
     Gender getGender() const override;
     const std::string& getCity() const override;
     const std::string& getStreet() const override;
-    std::optional<size_t> getIncome() const override { return income_; }
+    std::optional<size_t> getIncome() const override;
 
 private:
     std::string firstName_;
@@ -35,5 +33,4 @@ private:
     Gender gender_ = Gender::Undefined;
     Address address_;
     size_t income_{};
-    size_t indexNumber_{};
 };
